pyproject.toml
[project]
name = "news-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.11"
dependencies = [
    "advertools>=0.17.1",
    "autoflake>=2.3.1",
    "black>=25.9.0",
    "google-genai>=1.45.0",
    "isort>=7.0.0",
    "jupyter>=1.1.1",
    "langchain-community>=0.4",
    "langchain-core>=1.0.0",
    "langchain-tavily>=0.2.12",
    "openai>=2.5.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "trafilatura>=2.0.0",
]
